+++ src/win32/Srb2win-vc10.vcxproj.filters
@@ -255,6 +255,9 @@
     <ClCompile Include="..\lua_skinlib.c">
       <Filter>LUA</Filter>
     </ClCompile>
+    <ClCompile Include="..\m_aatree.c">
+      <Filter>M_Misc</Filter>
+    </ClCompile>
     <ClCompile Include="..\m_anigif.c">
       <Filter>M_Misc</Filter>
     </ClCompile>
@@ -660,6 +663,9 @@
       <Filter>LUA</Filter>
     </ClInclude>
     <ClInclude Include="..\md5.h">
+      <Filter>M_Misc</Filter>
+    </ClInclude>
+    <ClInclude Include="..\m_aatree.h">
       <Filter>M_Misc</Filter>
     </ClInclude>
     <ClInclude Include="..\m_anigif.h">
